There was a bug in the _paddr usage. It was actually a virtual address. We need only to store the virtual address returned from malloc. virt_to_phys is the function translating the virtual mem into physical.
In addition, I founf a design error in the virtio PFN code. It will only
affect us on > 44 bit ram..

+++ drivers/virtio.cc
@@ -101,7 +101,9 @@ namespace virtio {
             _queues[queuesel++] = queue;
 
             // Tell host about pfn
-            pci_conf_write(VIRTIO_PCI_QUEUE_PFN, (u32)((u64)queue->get_paddr() >> VIRTIO_PCI_QUEUE_ADDR_SHIFT));
+            // TODO: Yak, this is a bug in the design, on large memory we'll have PFNs > 32 bit
+            // Dor to notify Rusty
+            pci_conf_write(VIRTIO_PCI_QUEUE_PFN, (u32)(queue->get_paddr() >> VIRTIO_PCI_QUEUE_ADDR_SHIFT));
 
             // Debug print
             debug(fmt("Queue[%d] -> size %d, paddr %x") % (queuesel-1) % qsize % queue->get_paddr());
